requirements.txt
python-kucoin
requests
boto3
numpy<2.0.0
pandas
pandas-ta
matplotlib
seaborn
python-dotenv
openpyxl
psutil
pytest
pytest-cov
pytest-mock
black
flake8
mypy
bandit
safety
